requirements.txt
pywin32>=306
pandas>=2.2.0
flask>=3.0.0
flask-socketio>=5.3.0
python-dateutil>=2.9.0
project-x-py>=3.5.9
python-dotenv>=1.0.0
httpx[http2]>=0.28.0
polars>=1.20.0,<1.31.0
tzdata>=2024.1
claude-code-sdk>=0.1.0
psutil>=5.9.0
eventlet>=0.35.0
apscheduler>=3.10.0
